requirements.txt
mcp[cli]
python-dotenv
selenium
webdriver-manager
fastapi
uvicorn[standard]
pydantic